+++ packages/orq-rc/src/lib/config.ts
@@ -58,7 +58,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "2.0",
-  sdkVersion: "4.11.0-rc.9",
+  sdkVersion: "4.11.0-rc.10",
   genVersion: "2.899.1",
-  userAgent: "speakeasy-sdk/typescript 4.11.0-rc.9 2.899.1 2.0 @orq-ai/node",
+  userAgent: "speakeasy-sdk/typescript 4.11.0-rc.10 2.899.1 2.0 @orq-ai/node",
 } as const;